Fetch image details (#78)
* WIP

* Add labels fetching and showing

* Fetch labels and add styling

* Add key to li
